srvbase_qcc/requirements.txt
tensorflow-gpu==1.14.0
tensorflow==1.14.0
tensorpack==0.9.8
keras==2.1.4
pycocotools
opencv-python
Pillow
scikit-image
scikit-learn
Flask
flask_bootstrap
flask_cors
werkzeug
numpy
gast==0.2.2